sms/requirements.txt
asgiref==3.3.4
attrs==21.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coverage==5.5
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-allauth==0.44.0
django-autoslug==1.9.8
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-model-utils==4.1.1
django-phonenumber-field==5.1.0
django-widget-tweaks==1.4.8
flake8==3.9.1
flake8-isort==4.0.0
idna==2.10
iniconfig==1.1.1
isort==5.8.0
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.9
phonenumbers==8.12.22
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.11.1
pytest-cover==3.0.0
pytest-coverage==0.0
pytest-django==4.2.0
python-dotenv==0.17.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
testfixtures==6.17.1
toml==0.10.2
urllib3==1.26.4
whitenoise==5.2.0
